+++ packages/web/src/features/activity-report/components/executive/ActivityReportStatistic.tsx
