+++ packages/web/src/features/activity-report/frames/executive/ExecutiveActivityReportFrame.tsx
@@ -5,13 +5,12 @@ import AsyncBoundary from "@sparcs-clubs/web/common/components/AsyncBoundary";
 import Button from "@sparcs-clubs/web/common/components/Button";
 import FlexWrapper from "@sparcs-clubs/web/common/components/FlexWrapper";
 import SearchInput from "@sparcs-clubs/web/common/components/SearchInput";
-
-import ActivityReportStatistic from "../components/ActivityReportStatistic";
-import ChargedChangeClubModalContent from "../components/ChargedChangeClubModalContent";
-import { ChargedChangeClubProps } from "../components/ChargedChangeClubModalTable";
-import ExecutiveActivityChargedTable from "../components/ExecutiveActivityChargedTable";
-import ExecutiveActivityClubTable from "../components/ExecutiveActivityClubTable";
-import useGetExecutiveActivities from "../services/useGetExecutiveActivities";
+import ActivityReportStatistic from "@sparcs-clubs/web/features/activity-report/components/executive/ActivityReportStatistic";
+import ChargedChangeClubModalContent from "@sparcs-clubs/web/features/activity-report/components/executive/ChargedChangeClubModalContent";
+import { ChargedChangeClubProps } from "@sparcs-clubs/web/features/activity-report/components/executive/ChargedChangeClubModalTable";
+import ExecutiveActivityChargedTable from "@sparcs-clubs/web/features/activity-report/components/executive/ExecutiveActivityChargedTable";
+import ExecutiveActivityClubTable from "@sparcs-clubs/web/features/activity-report/components/executive/ExecutiveActivityClubTable";
+import useGetExecutiveActivities from "@sparcs-clubs/web/features/activity-report/services/executive/useGetExecutiveActivities";
 
 const ExecutiveActivityReportFrame = () => {
   const [isClubView, setIsClubView] = useState<boolean>(
